Rfactor: Created UserTx (transactions) and moved 2 hooks inside
