+++ app/src/hooks/MarketData/useTokens.tsx
@@ -3,8 +3,8 @@ import configs from '@makerdao/multicall/src/addresses.json'
 import axios from 'axios'
 import { useState } from 'react'
 
-import { DAI_TO_XDAI_TOKEN_BRIDGE_ADDRESS, OMNI_BRIDGE_MAINNET_ADDRESS } from '../common/constants'
-import { ConnectedWeb3Context } from '../contexts'
+import { DAI_TO_XDAI_TOKEN_BRIDGE_ADDRESS, OMNI_BRIDGE_MAINNET_ADDRESS } from '../../common/constants'
+import { ConnectedWeb3Context } from '../../contexts'
 import {
   getGraphUris,
   getInfuraUrl,
@@ -13,10 +13,10 @@ import {
   getTokensByNetwork,
   networkNames,
   pseudoNativeAssetAddress,
-} from '../util/networks'
-import { getImageUrl } from '../util/token'
-import { isObjectEqual } from '../util/tools'
-import { Token } from '../util/types'
+} from '../../util/networks'
+import { getImageUrl } from '../../util/token'
+import { isObjectEqual } from '../../util/tools'
+import { Token } from '../../util/types'
 
 const query = `
   query GetTokenList($listId: String!) {
