+++ app/src/hooks/index.ts
@@ -1,10 +1,10 @@
-export * from './useBalance'
+export * from './UserTx/useBalance'
 export { use24hsVolume } from './MarketData/use24hsVolume'
 export { useAsyncDerivedValue } from './useAsyncDerivedValue'
 export { useBlockchainMarketMakerData } from './MarketData/useBlockchainMarketMakerData'
 export { useCheckContractExists } from './useCheckContractExists'
 export { useCollateral } from './useCollateral'
-export { useCollateralBalance } from './useCollateralBalance'
+export { useCollateralBalance } from './UserTx/useCollateralBalance'
 export { useCompoundService } from './useCompoundService'
 export { useContracts } from './useContracts'
 export { useCpkAllowance } from './Cpk/useCpkAllowance'
@@ -12,13 +12,13 @@ export { useCpkProxy } from './Cpk/useCpkProxy'
 export { useDetectAdblocker } from './useDetectAdblocker'
 export { useDocumentDescription } from './useDocumentDescription'
 export { useDocumentTitle } from './useDocumentTitle'
-export { useFundingBalance } from './useFundingBalance'
+export { useFundingBalance } from './MarketData/useFundingBalance'
 export { useGraphMarketMakerData } from './Graph/useGraphMarketMakerData'
 export { useGraphMarketUserTxData } from './Graph/useGraphMarketUserTxData'
 export { useGraphMeta } from './Graph/useGraphMeta'
 export { useMarketMakerData } from './MarketData/useMarketMakerData'
 export { usePolling } from './usePolling'
-export { useTokens } from './useTokens'
+export { useTokens } from './MarketData/useTokens'
 export { useRealityLink } from './useRealityLink'
 export { useSymbol } from './useSymbol'
 export { useMultipleQueries } from './useMultipleQueries'
